+++ src/main/java/org/plumelib/util/Pair.java
@@ -7,12 +7,12 @@ import org.checkerframework.dataflow.qual.SideEffectFree;
 
 /**
  * Mutable pair class: type-safely holds two objects of possibly-different types. Defines
- * toString(), equals(), and hashCode() routines. If you don't need these features, or if you want
- * an immutable pair, you are better off defining your own tiny class.
+ * toString(), equals(), and hashCode() routines. If you don't need these features, you are better
+ * off defining your own tiny class with meaningful field names.
  *
  * @param <T1> the type of the first element of the pair
  * @param <T2> the type of the second element of the pair
- * @deprecated use MPair instead
+ * @deprecated use {@link MPair} instead, or {@link IPair} for an immutable pair
  */
 @Deprecated // 2023-05-20
 public class Pair<T1 extends @Nullable Object, T2 extends @Nullable Object> {
@@ -68,7 +68,6 @@ public class Pair<T1 extends @Nullable Object, T2 extends @Nullable Object> {
   // If fields a and b were made final, then the hashcode could be cached.
   // (And if they aren't final, it's a bit odd to be calling hashCode.)
   // But then the class would not be useful for mutable pairs.
-  @SuppressWarnings("signedness:override.receiver") // temporary
   @Override
   @Pure
   public int hashCode(@GuardSatisfied Pair<T1, T2> this) {