+++ src/main/java/org/plumelib/util/StringsPlume.java
@@ -3,7 +3,6 @@
 package org.plumelib.util;
 
 import java.io.Serializable;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Comparator;
 import java.util.Iterator;
@@ -783,7 +782,7 @@ public final class StringsPlume {
    *
    * @param s string to truncate or pad
    * @param length goal length
-   * @return s truncated or padded to length characters
+   * @return {@code s} truncated or padded to {@code length} characters
    */
   @SuppressWarnings({
     "allcheckers:purity.not.sideeffectfree.call", // side effect to local state
@@ -803,23 +802,24 @@ public final class StringsPlume {
   }
 
   /**
-   * Returns a string of the specified length, truncated if necessary, and padded with spaces to the
-   * right if necessary.
+   * Returns a string of the specified length, truncated if necessary, and padded with the given
+   * character to the right if necessary.
    *
    * @param s string to truncate or pad
    * @param length goal length
-   * @return s truncated or padded to length characters
+   * @param c character to use for padding
+   * @return {@code s} truncated or padded to {@code length} characters
    */
   @SuppressWarnings({
     "allcheckers:purity.not.sideeffectfree.call", // side effect to local state
     "lock:method.guarantee.violated" // side effect to local state
   })
   @SideEffectFree
-  public static String rpad(String s, @NonNegative int length) {
+  public static String rpad(String s, @NonNegative int length, char c) {
     if (s.length() < length) {
       StringBuilder buf = new StringBuilder(s);
       for (int i = s.length(); i < length; i++) {
-        buf.append(' ');
+        buf.append(c);
       }
       return buf.toString();
     } else {
@@ -828,11 +828,24 @@ public final class StringsPlume {
   }
 
   /**
+   * Returns a string of the specified length, truncated if necessary, and padded with spaces to the
+   * right if necessary.
+   *
+   * @param s string to truncate or pad
+   * @param length goal length
+   * @return {@code s} truncated or padded to {@code length} characters
+   */
+  @SideEffectFree
+  public static String rpad(String s, @NonNegative int length) {
+    return rpad(s, length, ' ');
+  }
+
+  /**
    * Converts the int to a String, then formats it using {@link #rpad(String,int)}.
    *
    * @param num int whose string representation to truncate or pad
    * @param length goal length
-   * @return a string representation of num truncated or padded to length characters
+   * @return a string representation of {@code num} truncated or padded to {@code length} characters
    */
   @SideEffectFree
   public static String rpad(int num, @NonNegative int length) {
@@ -844,7 +857,7 @@ public final class StringsPlume {
    *
    * @param num double whose string representation to truncate or pad
    * @param length goal length
-   * @return a string representation of num truncated or padded to length characters
+   * @return a string representation of {@code num} truncated or padded to {@code length} characters
    */
   @SideEffectFree
   public static String rpad(double num, @NonNegative int length) {
@@ -980,7 +993,7 @@ public final class StringsPlume {
     "lock:method.guarantee.violated" // side effect to local state
   })
   @SideEffectFree
-  public static ArrayList<Object> tokens(String str, String delim, boolean returnDelims) {
+  public static List<Object> tokens(String str, String delim, boolean returnDelims) {
     return CollectionsPlume.makeArrayList(new StringTokenizer(str, delim, returnDelims));
   }
 
@@ -997,7 +1010,7 @@ public final class StringsPlume {
     "lock:method.guarantee.violated" // side effect to local state
   })
   @SideEffectFree
-  public static ArrayList<Object> tokens(String str, String delim) {
+  public static List<Object> tokens(String str, String delim) {
     return CollectionsPlume.makeArrayList(new StringTokenizer(str, delim));
   }
 
@@ -1013,7 +1026,7 @@ public final class StringsPlume {
     "lock:method.guarantee.violated" // side effect to local state
   })
   @SideEffectFree
-  public static ArrayList<Object> tokens(String str) {
+  public static List<Object> tokens(String str) {
     return CollectionsPlume.makeArrayList(new StringTokenizer(str));
   }