+++ src/main/java/org/plumelib/util/ToStringComparator.java
@@ -4,6 +4,9 @@ import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Objects;
+import org.checkerframework.checker.mustcall.qual.MustCallUnknown;
+import org.checkerframework.checker.nullness.qual.Nullable;
+import org.checkerframework.checker.signedness.qual.Signed;
 
 /**
  * A comparator that orders values based on the lexicographic ordering of their toString().
@@ -21,7 +24,7 @@ public class ToStringComparator implements Comparator<Object> {
   private ToStringComparator() {}
 
   @Override
-  public int compare(Object o1, Object o2) {
+  public int compare(@MustCallUnknown Object o1, @MustCallUnknown Object o2) {
     return Objects.toString(o1).compareTo(Objects.toString(o2));
   }
 
@@ -32,12 +35,8 @@ public class ToStringComparator implements Comparator<Object> {
    * @param in a set of elements
    * @return the elements, sorted according to {@code toString()}
    */
-  @SuppressWarnings({
-    "signedness:argument", // temporary
-    "nullness:argument", // Comparator should be @Contravariant.
-    "mustcall:argument" // not sure; Java generics inference?
-  })
-  public static <T> List<T> sorted(Iterable<T> in) {
+  @SuppressWarnings("nullness:argument") // Comparator should be @Contravariant.
+  public static <T extends @Nullable @Signed Object> List<T> sorted(Iterable<T> in) {
     List<T> result = new ArrayList<T>();
     for (T object : in) {
       result.add(object);