+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Plume-Util change log
 
-## 1.8.0 (2023-??-??)
+## 1.8.2
+- `StringsPlume`: add `rpad` that pads with an arbitrary character
+- `CollectionsPlume`: add an overload for `mapCapacity()`
+- TODO: FilesPlume changes.
+
+## 1.8.1 (2023-06-02)
+
+- `CollectionsPlume`: remove conflicting overrides of `deepCopy()`
+- `IPair` and `MPair`: rename fields from `a` and `b` to `first` and `second`
+- `MPair`: new methods `cloneElements()`, `deepCopy()`, `deepCopyFirst()`, `deepCopySecond()`
+- `StringsPlume`: add `rpad` that pads with an arbitrary character
+- `FilesPlume`: new `InputStream` methods `available()`, `isWhitespaceOnly()`, `readCodePoint()`
+
+## 1.8.0 (2023-06-02)
 
 - Deprecated `Pair` class for mutable pairs; use `MPair` instead.
   `OrderedPairIterator` uses `MPair` instead of `Pair`.
@@ -8,6 +21,7 @@
 - The `WeakIdentityPair` constructor is deprecated in favor of the `of()` method.
 - New interface `DeepCopyable`.
 - New methods in `UtilPlume`:
+   * `firstNonNull()`
 - New methods in `CollectionsPlume`:
    * for `Collection`s: `cloneElements()`, `deepCopy()`
    * for `Map`s: `cloneElements()`, `cloneValues()`, `createLruCache()`